refactor(server): unify entrypoint via createApp factory
Replaces server/railway.mjs and server/replit.mjs (~95% duplicated, with
divergences) with a single server/index.js + server/app.js + server/config.js.

Runtime mode is selected by SERVER_MODE=railway|replit (or auto-detected
from REPLIT_DOMAINS), which drives port, SPA static serving, trust-proxy,
and helmet CSP strictness. Route registration is unified so Replit now
also gets /api/push/* endpoints (previously silently missing) and Sentry
hooks when SENTRY_DSN is configured.

API contract, health probes, metrics, and auth mount points are unchanged.
