+++ package.json
@@ -3,8 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "start": "node server/railway.mjs",
-    "start:replit": "node server/replit.mjs",
+    "start": "node server/index.js",
+    "start:replit": "SERVER_MODE=replit node server/index.js",
     "dev": "vite --port 5173 --host 0.0.0.0",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
