+++ server/sentry.js
@@ -12,10 +12,10 @@ const dsn = process.env.SENTRY_DSN;
 // ВАЖЛИВО: ініціалізація робиться у module top-level, а не в окремій функції,
 // яку треба викликати. У ESM (`"type": "module"`) усі `import` хостяться і
 // оцінюються ДО виконання тіла модуля, тому якщо викликати `Sentry.init()` з
-// тіла `railway.mjs`, `express`/`http` уже будуть завантажені й
+// тіла `server/index.js`, `express`/`http` уже будуть завантажені й
 // OpenTelemetry-інструментація стане no-op.
 //
-// Рішення: ставимо `Sentry.init()` саме тут, а у `railway.mjs` цей файл
+// Рішення: ставимо `Sentry.init()` саме тут, а у `server/index.js` цей файл
 // імпортується ПЕРШИМ — завдяки depth-first evaluation ESM-імпортів тіло
 // `sentry.js` виконається до того, як станеться `import express`.
 if (dsn) {
@@ -67,7 +67,7 @@ if (dsn) {
 /**
  * Сумісний з попередньою сигнатурою no-op — ініт уже відбувся при імпорті.
  * Лишаємо функцію, щоб подальші міграції могли експортувати зі стану (перевірки
- * `initialized`), а поточні виклики в `railway.mjs` не потрібно видаляти.
+ * `initialized`), а зовнішні виклики не потрібно видаляти.
  */
 export function initSentry() {
   // Нічого — Sentry.init() виконався на рівні модуля.
